[Privatization] Move ChangeReadOnlyPropertyWithDefaultValueToConstantRector here
